Added extract_measure function
It allows the user to select a set of FRFs based on different substrings. The function can be used in two modes:
Exact = False (The name of the FRF only has to contain a single substring to be added to the new set.
Exact = True (The name of the FRF has to contain all given substrings.